requirements.txt
boto==2.48.0
dj-database-url==0.4.2
Django==1.10.7
django-markdownx==2.0.21
django-storages==1.6.5
gunicorn==19.7.1
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.1
py==1.4.34
pytest==3.2.0
whitenoise==3.3.0
